% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_author_profiles.R
\name{download_author_profiles}
\alias{download_author_profiles}
\title{Download Author Profiles}
\usage{
download_author_profiles(res)
}
\arguments{
\item{res}{Scopus search results, including \code{entries}, which is a list of publication details.
Each entry should have the fields \code{author}, where \code{author} is a list of authors.
You can use the list of authids instead of a Scopus search result.}
}
\value{
A data frame containing detailed author profiles.
}
\description{
The download_author_profiles() function retrieves detailed author profiles
from the existing Scopus Data
}
\examples{

# Example 1
res <- list(
  meta_data = list(
    terms = "artificial intelligence"
  ),
  entries = list(
    list(
      author = list(
        list( `authid` = "24776780000" ),
        list( `authid` = "57940702100" )
      )
    )
  )
)
author_profiles <- download_author_profiles( res )

}
