% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_scopus_data.R
\name{download_scopus_data}
\alias{download_scopus_data}
\title{Download Scopus Data}
\usage{
download_scopus_data(
  terms,
  n = 10,
  search_type = "TITLE-ABS-KEY",
  other_fields = "",
  save_dir = "bibdata"
)
}
\arguments{
\item{terms}{A character vector of search terms (e.g., "machine learning" or c("machine learning", "algorithm")). These terms will be used to search for bibliographic records in the Scopus database.}

\item{n}{A integer specifying the search period. The default is "10".}

\item{search_type}{A character string specifying the search field. The default is "TITLE-ABS-KEY", which searches within titles, abstracts, and keywords. Other options can be specified as needed.}

\item{other_fields}{A character vector of additional search fields to include in the query.
Each entry should be a string representing a field condition, such as "AND AUTHOR-NAME(smith)".
These will be appended to the base query. Default is an empty string (""), meaning no additional fields.
Example: c("AND AUTHOR-NAME(smith)", "AND AFFILORG(University)").}

\item{save_dir}{A string specifying the directory where the results file will be saved.
If the directory does not exist, it will be created. Default is "bibdata".}
}
\value{
A data frame containing bibliographic records for the past 10 years (or n years, if you specify). The data frame includes various fields such as titles, authors, and other relevant metadata from the Scopus database. The specific columns in the data frame may vary depending on the search results.
}
\description{
The download_scopus_data() function retrieves bibliographic records from
the Scopus API for the past N years.
}
\examples{
# Example 1: Search for records related to "machine learning"
entries <- download_scopus_data( "journalist robot", search_type = "TITLE" )

# Example 2: Search for records related to multiple terms
terms <- c("machine learning", "algorithm")
entries <- download_scopus_data( terms, other_fields = "AND AUTHOR-NAME(matsumoto)" )
head(entries)

}
