% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_search_terms.R
\name{expand_search_terms}
\alias{expand_search_terms}
\title{Expand Search Terms}
\usage{
expand_search_terms(naive_terms)
}
\arguments{
\item{naive_terms}{A character vector of one or more initial search queries.}
}
\value{
A character vector containing the original queries along with expanded search terms.
}
\description{
The \code{expand_search_terms()} function takes one or more initial search terms and
returns a query (expanded set of terms) by including related or similar terms.
The expansion helps to capture more relevant results by broadening the search.
}
\examples{
expand_search_terms( c( "data analysis", "machine learning" ) )

}
