% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_search_query.R
\name{get_search_query}
\alias{get_search_query}
\title{Extract the search query from a given URL.}
\usage{
get_search_query(res)
}
\arguments{
\item{res}{A list containing Scopus search results, with a URL field in \code{get_statements}.}
}
\value{
A character string containing the search query.
}
\description{
This function extracts the search query string from a URL containing
Scopus API search parameters.
}
\examples{
res <- list(
  get_statements = list(
    url = "https://api.elsevier.com/content/search/scopus?query=TITLE-ABS-KEY\%20\%28\%22artificial\%20intelligence\%22\%20AND\%20\%22national\%20security\%22\%20OR\%20\%22machine\%20learning\%22\%29"
  )
)
#' get_search_query( res )

}
