% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_search_terms.R
\name{get_search_terms}
\alias{get_search_terms}
\title{Extract search terms from a Scopus search URL.}
\usage{
get_search_terms(res)
}
\arguments{
\item{res}{A list containing Scopus search results, with a URL field in \code{get_statements}.}
}
\value{
A list containing the search keywords and other query parameters.
}
\description{
This function extracts search keywords and other query parameters from the provided Scopus search URL
within the \code{res} object. The URL is obtained from \code{res$get_statements$url}.
}
\examples{
res <- list(
  get_statements = list(
    url = "https://api.elsevier.com/content/search/scopus?query=TITLE-ABS-KEY\%20\%28\%20\%22oryza\%20sativa\%22\%20\%29\%20\%20AND\%20PUBYEAR\%20\%3D\%202014&count=25&start=0&view=COMPLETE"
  )
)
get_search_terms(res)

}
