% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_authnum_trend.R
\name{plot_authnum_trend}
\alias{plot_authnum_trend}
\title{Plot Author Number Trend}
\usage{
plot_authnum_trend(res, type = "line")
}
\arguments{
\item{res}{Scopus search results, including \code{entries}, which is a list of publication details.
Each entry should have the fields \code{dc:title} and \code{author}, where \code{author} is a list of authors.}

\item{type}{A string specifying the type of plot. Options are "line" (default), "stack", and "ratio".
"line" produces a line plot, "stack" produces a stacked area plot, and "ratio" produces a 100\% stacked area plot.}
}
\value{
A ggplot2 object showing the trend analysis results over the years for each author number category.
}
\description{
The \code{plot_authnum_trend()} function analyses and visualizes
the annual publication trends for the number of authors
from Scopus search results.
You can specify \code{type} of the output from \code{line}, \code{stack} and \code{ratio}.
}
\examples{
# Create a sample `res` list to simulate Scopus API search results
res <- list(
  entries = list(
    list(
      `dc:title` = "A Study on Data Analysis",
      `prism:coverDate` = "2022-12-31",
      author = list(
        list( `authname` = "Author 1" ),
        list( `authname` = "Author 2" )
      )
    ),
    list(
      `dc:title` = "Advances in Machine Learning",
      `prism:coverDate` = "2021-11-30",
      author = list(
        list( `authname` = "Author 1" ),
        list( `authname` = "Author 2" ),
        list( `authname` = "Author 3" ),
        list( `authname` = "Author 4" ),
        list( `authname` = "Author 5" )
      )
    ),
    list(
      `dc:title` = "Research in Quantum Computing",
      `prism:coverDate` = "2020-10-15",
      author = list(
        list( `authname` = "Author 1" )
      )
    )
  )
)

# Example usage of the function with the sample data
plot_authnum_trend( res, type = "stack" )

}
