% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_country_pairs_collaboration_ratio.R
\name{plot_country_pairs_collaboration_ratio}
\alias{plot_country_pairs_collaboration_ratio}
\title{Plot Country Pairs Collaboration Ratio}
\usage{
plot_country_pairs_collaboration_ratio(
  res,
  n = 5,
  target_country = NULL,
  type = "line"
)
}
\arguments{
\item{res}{List structure containing publication data}

\item{n}{Number of top country pairs to plot (default is 5)}

\item{target_country}{Optional; specific country to filter the pairs by (default is NULL)}

\item{type}{Type of a graph}
}
\value{
A plot object showing the trend of top country pairs by publication year
}
\description{
The \code{plot_country_pairs_collaboration_ratio()} function analyses and visualizes
the annual publication trends for country pairs with their collaboration ratio
from Scopus search results.
You can filter the result with a target country by specifying \code{target_country}
and choose the \code{type} of the output from \code{line} or \code{stack}.
}
