% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_affiliations.R
\name{rank_affiliations}
\alias{rank_affiliations}
\title{Generate a ranking of the top affiliation used in the provided Scopus search results.}
\usage{
rank_affiliations(res, n = 20, type = "country")
}
\arguments{
\item{res}{A list containing Scopus search results, including entries with \code{affiliation}.}

\item{n}{An integer specifying the number of top keywords to return. Defaults to 20.}

\item{type}{A character indicating what to rank. The value should be \code{country}, \code{city}, or \code{institution}. Default is \code{country}.}
}
\value{
A tibble with columns
\code{rank} (integer, rank),
\code{institution/country/city} (character), and
\code{n} (integer, occurrence count).
}
\description{
The \code{rank_affiliations()} function generates a ranking of the top affiliations
from the provided Scopus search results,
based on the number of publications and total citations.
By choosing the \code{type = "city"} or \code{type = "institution"} option,
the function can provide the ranking of cities or institutions instead of countries.
The output is a tibble with the following columns:
\code{rank} rank of the affiliation,
\code{country/city/institution} type of the affiliation, and
\code{n} number of publications.
}
\examples{
# Load a sample Scopus search result object (res)
# Note: This is a placeholder example. In practice, you would load or obtain a real 'res' object.
res <- list(
  entries = list(
    list(
      affiliation = list(
        list(
          `affiliation-country` = c("USA", "Canada", "Germany"),
          `affiliation-city` = c("New York", "Toronto", "Berlin"),
          `affilname` = c("Johns Hopkins University")
        )
      )
    ),
    list(
      affiliation = list(
        list(
          `affiliation-country` = c("USA", "France", "Germany"),
          `affiliation-city` = c("San Francisco", "Paris", "Berlin"),
          `affilname` = c("University of California")
        )
      )
    )
  )
)

# Generate a ranking of the top 20 countries
rank_affiliations( res )

# Generate a ranking of the top 20 cities
rank_affiliations( res, n = 20, type = "city")

}
