% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_keywords.R
\name{rank_keywords}
\alias{rank_keywords}
\title{Generate a ranking of the top keywords used in the provided Scopus search results.}
\usage{
rank_keywords(res, n = 20, abst = FALSE)
}
\arguments{
\item{res}{A list containing Scopus search results, including entries with \code{authkeywords} and \code{dc:description}.}

\item{n}{An integer specifying the number of top keywords to return. Defaults to 20.}

\item{abst}{A logical value indicating whether to include keywords from the abstracts. Default is \code{FALSE}.}
}
\value{
A tibble with columns
\code{rank} (integer, rank),
\code{keyword} (character, keyword), and
\code{n} (integer, occurrence count).
}
\description{
The \code{rank_keywords()} function generates a ranking of the top keywords
from the provided Scopus search results,
based on the number of publications and total citations.
By setting the \code{abst = TRUE} option, the function can include
not only the keywords tagged by the authors
but also the abstracts of the papers in the ranking.
The output is a tibble with the following columns:
\code{rank} rank of the keyword,
\code{keyword} the keyword, and
\code{n} number of publications with the keyword.
}
\examples{
res <- list(
  entries = list(
    list( authkeywords = "machine learning | AI | deep learning",
          `dc:description` = "This is an abstract about AI." ),
    list( authkeywords = "machine learning | data mining",
          `dc:description` = "Abstract about data mining and machine learning." )
  )
)

rank_keywords( res, abst = TRUE )

}
