% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_sponsors.R
\name{rank_sponsors}
\alias{rank_sponsors}
\title{Rank Sponsors by Number of Publications and Citations}
\usage{
rank_sponsors(res, n = 10)
}
\arguments{
\item{res}{A list containing Scopus search results.}

\item{n}{An integer specifying the number of top sponsors to return. Defaults to 20.}
}
\value{
A tibble with the ranking of sponsors.
}
\description{
This function ranks sponsors based on the number of publications and total citations.
The output is a tibble with the following columns:
\itemize{
\item \code{rank}: Rank of the sponsor.
\item \code{sponsor}: Sponsor's name with abbreviation.
\item \code{n_publications}: Number of publications.
\item \code{n_citations}: Total number of citations.
\item \code{cite_per_pub}: Average number of citations per publication.
\item \code{countries_display}: Comma-separated list of countries where the sponsor is active.
\item \code{countries}: List of countries where the sponsor is active.
}
}
\examples{
res <- list(
  entries = list(
    list(`fund-sponsor` = "Youth Innovation Promotion Association",
         `fund-acr` = "YIPA",
         affiliation = list(list(`affiliation-country` = c("Country A", "Country B"))),
         `citedby-count` = 10),
    list(`fund-sponsor` = "National Science Foundation",
         `fund-acr` = "NSF",
         affiliation = list(list(`affiliation-country` = c("Country C"))),
         `citedby-count` = 5)
  )
)
rank_sponsors(res, n = 20)

}
