% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna_countries.R
\name{sna_countries}
\alias{sna_countries}
\title{Social Network Analysis - Network of Country Co-occurrences}
\usage{
sna_countries(res)
}
\arguments{
\item{res}{A list containing Scopus search results, which includes:
\itemize{
\item \code{entries}: A list of publication entries where each entry contains:
\itemize{
\item \code{authkeywords}: Keywords related to the publication, separated by \code{|}.
\item \code{affiliation}: A list of affiliation details, including:
\itemize{
\item \code{affiliation-country}: The country associated with the publication's affiliation.
}
}
}}
}
\value{
A \code{ggraph} object showing the network of countries based on their co-occurrence in research topics.
The network visualizes countries as nodes and the strength of their connections (based on shared keywords)
as edges between these nodes. Nodes are sized according to their importance in the network, and edges
are colored based on their weight.
}
\description{
This function analyzes co-occurrence data between countries based on shared keywords
and creates a network graph visualizing the relationships between countries.
The network is constructed based on the frequency of shared keywords between countries.
The resulting graph can help in understanding how countries are connected through common research topics.
}
\examples{
res <- list(
  entries = list(
    list(
      authkeywords = "Artificial Intelligence|Machine Learning",
      affiliation = list(
        list( `affiliation-country` = "USA" )
      )
    ),
    list(
      authkeywords = "Artificial Intelligence|Data Science",
      affiliation = list(
        list( `affiliation-country` = "UK" )
      )
    ),
    list(
      authkeywords = "Machine Learning|Data Science",
      affiliation = list(
        list( `affiliation-country` = "Canada")
      )
    ),
    list(
      authkeywords = "Artificial Intelligence|Machine Learning|Data Science",
      affiliation = list(
        list( `affiliation-country` = "Germany")
      )
    )
  )
)

network_graph <- sna_countries( res )
print( network_graph )

}
