% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_search_terms.R
\name{get_search_terms}
\alias{get_search_terms}
\title{Extract search terms from a Scopus search URL.}
\usage{
get_search_terms(res)
}
\arguments{
\item{res}{A list containing Scopus search results, with a URL field in \code{get_statements}.}
}
\value{
A list of search keywords extracted from the URL.
}
\description{
This function extracts search keywords from the provided Scopus search URL
within the \code{res} object. The URL is obtained from \code{res$get_statements$url}.
}
\examples{
res <- list(
  get_statements = list(
    url = "https://api.elsevier.com/content/search/scopus?query=TITLE-ABS-KEY\%20\%28\%22artificial\%20intelligence\%22\%20AND\%20\%22national\%20security\%22\%20OR\%20\%22machine\%20learning\%22\%29"
  )
)
#' get_search_terms( res )

}
