% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_country_pairs_trend.R
\name{plot_country_pairs_trend}
\alias{plot_country_pairs_trend}
\title{Plot Country Pairs Trend}
\usage{
plot_country_pairs_trend(res, n = 5, target_country = NULL)
}
\arguments{
\item{res}{List structure containing publication data}

\item{n}{Number of top country pairs to plot (default is 5)}

\item{target_country}{Optional; specific country to filter the pairs by (default is NULL)}
}
\value{
A plot object showing the trend of top country pairs by publication year
}
\description{
The \code{plot_country_pairs_trend()} function analyses and visualizes
the annual publication trends for country pairs
from Scopus search results.
You can filter the result with a target country by specifying \code{target_country}.
}
