% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_sponsors.R
\name{rank_sponsors}
\alias{rank_sponsors}
\title{Rank Sponsors by Number of Publications and Citations}
\usage{
rank_sponsors(res, n = 10)
}
\arguments{
\item{res}{A list containing Scopus search results.}

\item{n}{An integer specifying the number of top sponsors to return. Defaults to 20.}
}
\value{
A tibble with the ranking of sponsors.
}
\description{
The \code{rank_sponsors()} function generates a ranking of the top sponsors
from the provided Scopus search results,
based on the number of publications and total citations.
The output is a tibble with the following columns:
\code{rank} rank of the sponsor,
\code{sponsor} sponsor's name with abbreviation,
\code{n_publications} number of publications,
\code{n_citations} total number of citations,
\code{cite_per_pub} average number of citations per publication,
\code{countries_display} comma-separated list of countries where the sponsor is active, and
\code{countries} list of countries where the sponsor is active.
}
\examples{
res <- list(
  entries = list(
    list(`fund-sponsor` = "Youth Innovation Promotion Association",
         `fund-acr` = "YIPA",
         affiliation = list(list(`affiliation-country` = c("Country A", "Country B"))),
         `citedby-count` = 10),
    list(`fund-sponsor` = "National Science Foundation",
         `fund-acr` = "NSF",
         affiliation = list(list(`affiliation-country` = c("Country C"))),
         `citedby-count` = 5)
  )
)
rank_sponsors(res, n = 20)

}
